uploaded database file
I did this because when I checked out the project to a new directory the application showed none of the data  had created.
Not sure how to deal with this in the future as it seems a lot to store on git hub. I wonder if I need to export the database data in flat files
and import each time I generate a new project. Perhaps I'm over thingg it